refactor: move shared logic to context api
+ Move shared logic to context provider API
+ Misc. tweaks
