+++ packages/app/src/components/live-editor.js
@@ -18,8 +18,8 @@ const LivePreviewWrapper = styled('div')`
   margin: auto;
   overflow: hidden;
 
-  ${({ isEditor }) =>
-    isEditor &&
+  ${({ shadow }) =>
+    shadow &&
     `
   box-shadow: rgba(0, 0, 0, 0.12) 0px 5px 10px 0px;`}
 `
