+++ packages/app/src/context/overlay-context.js
@@ -1,0 +1,23 @@
+import { useCallback, useState } from 'react'
+import defer from 'tickedoff'
+
+const overlayContext = (onShow) => {
+  const [isOverlay, setOverlay] = useState('')
+
+  const showOverlay = useCallback(
+    (state) => () => {
+      if (onShow) {
+        onShow()
+      }
+
+      defer(() => setOverlay(state))
+    },
+    [onShow]
+  )
+
+  const hideOverlay = useCallback(() => setOverlay(''), [])
+
+  return { isOverlay, showOverlay, hideOverlay }
+}
+
+export default overlayContext
